+++ lib/gs_form/model/data_model/spinner_data_model.dart
@@ -4,7 +4,6 @@ class SpinnerDataModel {
   bool? isSelected;
   dynamic data;
 
-
-  SpinnerDataModel({required this.name, required this.id, this.data, bool ?isSelected})
-      : isSelected = isSelected ?? false ;
+  SpinnerDataModel({required this.name, required this.id, this.data, bool? isSelected})
+      : isSelected = isSelected ?? false;
 }